ai_functions/requirements.txt
torch>=1.13.0
numpy>=1.23.0
gymnasium>=0.28.0
psycopg2-binary>=2.9.0
python-dotenv>=0.21.0